requirements.txt
asgiref==3.8.1
bitarray==3.4.2
contourpy==1.3.2
cycler==0.12.1
Django==5.2.1
django-jazzmin==3.0.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flake8==7.3.0
fonttools==4.58.4
kiwisolver==1.4.8
matplotlib==3.10.3
mccabe==0.7.0
numpy==2.3.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
